requirements_clean.txt
# requirements_clean.txt - Railway Compatible + OpenAI
# Versione pulita per deployment incrementale

# ===== CORE FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.5.0

# ===== DATABASE & ASYNC =====
aiosqlite==0.19.0

# ===== UTILITIES =====
python-dotenv==1.0.0
psutil==5.9.0
typing-extensions==4.8.0
httpx==0.25.0

# ===== OPENAI INTEGRATION (NEW!) =====
openai==1.3.0

# ===== FUTURE ROADMAP (commentate per ora) =====
# Aggiungi gradualmente quando Railway è stabile:

# FASE 2: LangChain RAG
# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-community==0.0.10

# FASE 3: Vector Database  
# chromadb==0.4.15
# langchain-chroma==0.1.0

# FASE 4: Document Processing
# pymupdf==1.23.0        # Quando Railway supporta C++ compilation
# pytesseract==0.3.10    # Quando Tesseract è disponibile
# Pillow==10.1.0
# unstructured==0.11.0
# python-docx==1.1.0

# FASE 5: NLP Advanced
# spacy==3.7.0           # Quando memory usage è acceptable

# ===== NOTE =====
# FASE 1: Solo OpenAI - Sistema minimal funzionante
# FASE 2-5: Aggiungi incrementalmente per sistema completo